Fix: filepaths and meson build system for refactor
