+++ tests/test_tools.c
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-#include "../tools.h"
+#include "../src/tools.h"
 #include "utils.h"
 #include <errno.h>
 #include <math.h>
